:recycle: Make clear that flatten_dict is not a user facing function by renaming it _flatten_dict
