+++ kedro_mlflow/framework/hooks/utils.py
@@ -1,3 +1,5 @@
+from typing import Dict
+
 from kedro_mlflow.framework.context.mlflow_context import get_mlflow_config
 
 
@@ -12,3 +14,16 @@ def _assert_mlflow_enabled(pipeline_name: str) -> bool:
         return False
 
     return True
+
+
+def _flatten_dict(d: Dict, recursive: bool = True, sep: str = ".") -> Dict:
+    def expand(key, value):
+        if isinstance(value, dict):
+            new_value = _flatten_dict(value) if recursive else value
+            return [(f"{key}{sep}{k}", v) for k, v in new_value.items()]
+        else:
+            return [(f"{key}", value)]
+
+    items = [item for k, v in d.items() for item in expand(k, v)]
+
+    return dict(items)
